feat: update ceresdbproto and impl EventDispatch
