+++ server/cluster/shard.go
@@ -18,8 +18,6 @@ func (s *Shard) dropTableLocked(tableID uint64) {
 }
 
 type ShardTablesWithRole struct {
-	shardID   uint32
-	shardRole clusterpb.ShardRole
-	tables    []*Table
-	version   uint64
+	shard  *ShardInfo
+	tables []*Table
 }
