+++ server/coordinator/procedure/shard/fsm.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/CeresDB/ceresdbproto/pkg/clusterpb"
 	"github.com/CeresDB/ceresmeta/server/cluster"
-	"github.com/CeresDB/ceresmeta/server/coordinator/procedure/dispatch"
+	"github.com/CeresDB/ceresmeta/server/coordinator/dispatch"
 	"github.com/looplab/fsm"
 	"github.com/pkg/errors"
 )
@@ -57,18 +57,18 @@ var (
 )
 
 type LeaderCallbackRequest struct {
-	Ctx            context.Context
-	Cluster        *cluster.Cluster
-	ActionDispatch dispatch.ActionDispatch
+	Ctx           context.Context
+	Cluster       *cluster.Cluster
+	EventDispatch dispatch.EventDispatch
 
 	Node    string
 	ShardID uint32
 }
 
 type FollowerCallbackRequest struct {
-	Ctx            context.Context
-	Cluster        *cluster.Cluster
-	ActionDispatch dispatch.ActionDispatch
+	Ctx           context.Context
+	Cluster       *cluster.Cluster
+	EventDispatch dispatch.EventDispatch
 
 	Node    string
 	ShardID uint32
@@ -121,12 +121,12 @@ func transferFollowerCallback(event *fsm.Event) {
 	request := event.Args[0].(*LeaderCallbackRequest)
 	ctx := request.Ctx
 	c := request.Cluster
-	actionDispatch := request.ActionDispatch
+	EventDispatch := request.EventDispatch
 	node := request.Node
 	shardID := request.ShardID
 
-	action := dispatch.CloseShardAction{ShardIDs: []uint32{shardID}}
-	if err := actionDispatch.CloseShards(ctx, node, action); err != nil {
+	action := &dispatch.CloseShardRequest{ShardID: shardID}
+	if err := EventDispatch.CloseShard(ctx, node, action); err != nil {
 		event.Cancel(errors.Wrap(err, EventPrepareTransferFollower))
 	}
 
@@ -139,13 +139,13 @@ func transferFollowerFailedCallback(event *fsm.Event) {
 	request := event.Args[0].(*LeaderCallbackRequest)
 	ctx := request.Ctx
 	c := request.Cluster
-	actionDispatch := request.ActionDispatch
+	EventDispatch := request.EventDispatch
 	node := request.Node
 	shardID := request.ShardID
 
 	// Transfer failed, stop transfer and reset state.
-	action := dispatch.CloseShardAction{ShardIDs: []uint32{shardID}}
-	if err := actionDispatch.CloseShards(ctx, node, action); err != nil {
+	action := &dispatch.CloseShardRequest{ShardID: shardID}
+	if err := EventDispatch.CloseShard(ctx, node, action); err != nil {
 		event.Cancel(errors.Wrap(err, EventTransferFollowerFailed))
 	}
 
@@ -162,14 +162,19 @@ func prepareTransferLeaderCallback(_ *fsm.Event) {
 func transferLeaderCallback(event *fsm.Event) {
 	request := event.Args[0].(*FollowerCallbackRequest)
 	ctx := request.Ctx
-	actionDispatch := request.ActionDispatch
+	EventDispatch := request.EventDispatch
 	c := request.Cluster
 	node := request.Node
 	shardID := request.ShardID
 
 	// Send event to CeresDB, waiting for response
-	action := dispatch.OpenShardAction{ShardIDs: []uint32{shardID}}
-	if err := actionDispatch.OpenShards(ctx, node, action); err != nil {
+	// TODO: add shardInfo in FollowerCallbackRequest
+	action := &dispatch.OpenShardRequest{Shard: &cluster.ShardInfo{
+		ShardID:   shardID,
+		ShardRole: clusterpb.ShardRole_LEADER,
+		Version:   0,
+	}}
+	if err := EventDispatch.OpenShard(ctx, node, action); err != nil {
 		event.Cancel(errors.Wrap(err, EventTransferLeader))
 	}
 
@@ -181,14 +186,14 @@ func transferLeaderCallback(event *fsm.Event) {
 func transferLeaderFailed(event *fsm.Event) {
 	request := event.Args[0].(*FollowerCallbackRequest)
 	ctx := request.Ctx
-	actionDispatch := request.ActionDispatch
+	EventDispatch := request.EventDispatch
 	c := request.Cluster
 	node := request.Node
 	shardID := request.ShardID
 
 	// Transfer failed, stop transfer and reset state.
-	action := dispatch.CloseShardAction{ShardIDs: []uint32{shardID}}
-	if err := actionDispatch.CloseShards(ctx, node, action); err != nil {
+	action := &dispatch.CloseShardRequest{ShardID: shardID}
+	if err := EventDispatch.CloseShard(ctx, node, action); err != nil {
 		event.Cancel(errors.Wrap(err, EventTransferLeaderFailed))
 	}
